+++ lib/flutter_roster.dart
@@ -4,6 +4,14 @@
 
 library roster;
 
+export 'package:flutter_date_time_picker/src/models/date_box_base_theme.dart';
+export 'package:flutter_date_time_picker/src/models/date_box_current_theme.dart';
+export 'package:flutter_date_time_picker/src/models/date_box_disabled_theme.dart';
+export 'package:flutter_date_time_picker/src/models/date_box_highlight_theme.dart';
+export 'package:flutter_date_time_picker/src/models/date_box_selected_theme.dart';
+export 'package:flutter_date_time_picker/src/models/date_time_picker_bar_theme.dart';
+export 'package:flutter_date_time_picker/src/models/date_time_picker_theme.dart';
+
 export 'package:flutter_roster/src/models/roster_event.dart';
 export 'package:flutter_roster/src/models/roster_theme.dart';
 export 'package:flutter_roster/src/roster.dart';