+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [1.0.0] - 18 November 2022
+
+* Support for both horizontal and vertical variant
+* Adjustable size for the component
+
 ## [0.0.1] - 13 September 2022
 
 * Initial release.